Cleared up print statments, and put searchRoute into a namespace
